+++ leap_net/__init__.py
@@ -6,7 +6,7 @@
 # SPDX-License-Identifier: MPL-2.0
 # This file is part of leap_net, leap_net a keras implementation of the LEAP Net model.
 
-__version__ = "0.1.0"
+__version__ = "0.1.1"
 __all__ = []
 
 try: